+++ scripts/fresh-install-of-osx.sh
@@ -15,7 +15,7 @@
 ######################################################################################################################
 # Set DNS of 8.8.8.8 before proceeding (in some cases, for eg Jio Wifi, github doesn't resolve at all and times out) #
 ######################################################################################################################
-echo "==> Setting DNS for WiFi"
+echo '==> Setting DNS for WiFi'
 sudo networksetup -setdnsservers Wi-Fi 8.8.8.8
 
 #################################################################################################
@@ -27,16 +27,6 @@ if ! type warn &> /dev/null 2>&1; then
   FIRST_INSTALL=true source "${HOME}/.shellrc"
 else
   warn "skipping downloading and sourcing '${HOME}/.shellrc' since its already loaded"
-fi
-
-##################################
-# Install command line dev tools #
-##################################
-section_header "Installing xcode command-line tools"
-if ! is_directory "/Library/Developer/CommandLineTools/usr/bin"; then
-  reinstall_xcode_cmdline_tools
-else
-  warn "skipping installation of xcode command-line tools since its already present"
 fi
 
 ###############################
@@ -52,23 +42,33 @@ fi
 #####################
 # Turn on FileVault #
 #####################
-section_header "Verifying FileVault status"
+section_header 'Verifying FileVault status'
 if [[ "$(fdesetup isactive)" != "true" ]]; then
-  echo "$(red "FileVault is not turned on. Please encrypt your hard disk!")"
+  echo "$(red 'FileVault is not turned on. Please encrypt your hard disk!')"
   exit 1
+fi
+
+##################################
+# Install command line dev tools #
+##################################
+section_header 'Installing xcode command-line tools'
+if ! is_directory '/Library/Developer/CommandLineTools/usr/bin'; then
+  reinstall_xcode_cmdline_tools
+else
+  warn 'skipping installation of xcode command-line tools since its already present'
 fi
 
 #################################
 # Setup ssh scripts/directories #
 #################################
-section_header "Setting ssh config file permissions"
+section_header 'Setting ssh config file permissions'
 ensure_dir_exists_if_var_defined "${HOME}/.ssh"
 test -n "$(ls -A "${HOME}/.ssh")" && sudo chmod -R 600 "${HOME}"/.ssh/* || true
 
 #################################################################################
 # Ensure that some of the directories corresponding to the env vars are created #
 #################################################################################
-section_header "Creating directories defined by various env vars"
+section_header 'Creating directories defined by various env vars'
 ensure_dir_exists_if_var_defined "${DOTFILES_DIR}"
 ensure_dir_exists_if_var_defined "${PROJECTS_BASE_DIR}"
 ensure_dir_exists_if_var_defined "${PERSONAL_BIN_DIR}"
@@ -82,6 +82,7 @@ ensure_dir_exists_if_var_defined "${XDG_STATE_HOME}"
 ############################
 # Disable macos gatekeeper #
 ############################
+# section_header 'Disabling macos gatekeeper'
 # sudo spectl --master-disable
 
 #####################
@@ -90,6 +91,7 @@ ensure_dir_exists_if_var_defined "${XDG_STATE_HOME}"
 section_header "Installing oh-my-zsh into '$(yellow "${HOME}/.oh-my-zsh")'"
 if ! is_directory "${HOME}/.oh-my-zsh"; then
   sh -c "$(ZSH= curl -fsSL https://install.ohmyz.sh/)" "" --unattended
+  success "Successfully installed oh-my-zsh into '$(yellow "${HOME}/.oh-my-zsh")'"
 else
   warn "skipping installation of oh-my-zsh since '${HOME}/.oh-my-zsh' is already present"
 fi
@@ -98,14 +100,14 @@ fi
 # Install custom omz plugins #
 ##############################
 # Note: Some of these are available via brew, but enabling them will take an additional step and the only other benefit (of keeping them up-to-date using brew can still be achieved by updating the git repos directly)
-section_header "Installing custom omz plugins"
+section_header 'Installing custom omz plugins'
 ZSH_CUSTOM="${ZSH_CUSTOM:-${ZSH:-${HOME}/.oh-my-zsh}/custom}"
 ensure_dir_exists_if_var_defined "${ZSH_CUSTOM}/plugins"
 clone_if_not_present() {
-  target_folder="${ZSH_CUSTOM}/plugins/$(basename ${1})"
+  local target_folder="${ZSH_CUSTOM}/plugins/$(basename ${1})"
   if ! is_directory "${target_folder}"; then
     git clone -q --depth=1 "${1}" "${target_folder}"
-    success "Successfully cloned ${1} into ${target_folder}"
+    success "Successfully cloned oh-my-zsh plugin ${1} into ${target_folder}"
   else
     warn "skipping cloning of '$(basename "${1}")' since '${target_folder}' is already present"
   fi
@@ -125,6 +127,8 @@ if is_non_zero_string "${DOTFILES_DIR}" && ! is_git_repo "${DOTFILES_DIR}"; then
 
   # Note: Cloning with https since the ssh keys will not be present at this time
   git clone -q "https://github.com/${GH_USERNAME}/dotfiles" "${DOTFILES_DIR}"
+  success "Successfully cloned the dotfiles repo into ${DOTFILES_DIR}"
+
   git -C "${DOTFILES_DIR}" switch "${DOTFILES_BRANCH}"
   if [[ "$(git -C "${DOTFILES_DIR}" branch --show-current)" != "${DOTFILES_BRANCH}" ]]; then
     echo "$(red "'DOTFILES_BRANCH' env var is not equal to the branch that was checked out; something is wrong. Please correct before retrying!")"
@@ -148,8 +152,9 @@ if is_non_zero_string "${DOTFILES_DIR}" && ! is_git_repo "${DOTFILES_DIR}"; then
   if [ $? -ne 0 ]; then
     git -C "${DOTFILES_DIR}" remote add upstream "https://github.com/${UPSTREAM_GH_USERNAME}/dotfiles"
     git -C "${DOTFILES_DIR}" fetch --all
+    success 'Successfully set new upstream remote for the dotfiles repo'
   else
-    warn "skipping setting new upstream remote for the dotfiles repo"
+    warn 'skipping setting new upstream remote for the dotfiles repo'
   fi
 else
   # Load all zsh config files for PATH and other env vars to take effect
@@ -173,6 +178,7 @@ if ! command_exists brew; then
   chmod u+w "${HOMEBREW_PREFIX}"
 
   NONINTERACTIVE=1 bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+  success 'Successfully installed homebrew'
 
   eval "$(${HOMEBREW_PREFIX}/bin/brew shellenv)"
 else
@@ -192,57 +198,64 @@ replace_executable_if_exists_and_is_not_symlinked() {
   fi
 }
 
-section_header "Linking keybase for command-line invocation"
-if is_directory "/Applications/Keybase.app"; then
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/Keybase.app/Contents/SharedSupport/bin/keybase" "${HOMEBREW_PREFIX}/bin/keybase"
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/Keybase.app/Contents/SharedSupport/bin/git-remote-keybase" "${HOMEBREW_PREFIX}/bin/git-remote-keybase"
+section_header 'Linking keybase for command-line invocation'
+if is_directory '/Applications/Keybase.app'; then
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/Keybase.app/Contents/SharedSupport/bin/keybase' "${HOMEBREW_PREFIX}/bin/keybase"
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/Keybase.app/Contents/SharedSupport/bin/git-remote-keybase' "${HOMEBREW_PREFIX}/bin/git-remote-keybase"
+  success 'Successfully linked keybase into PATH'
 else
-  warn "skipping symlinking keybase for command-line invocation"
+  warn 'skipping symlinking keybase for command-line invocation'
 fi
 
-section_header "Linking VSCode/VSCodium for command-line invocation"
-if is_directory "/Applications/VSCodium - Insiders.app"; then
+section_header 'Linking VSCode/VSCodium for command-line invocation'
+if is_directory '/Applications/VSCodium - Insiders.app'; then
   # Symlink from the embedded executable for codium-insiders
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/VSCodium - Insiders.app/Contents/Resources/app/bin/codium-insiders" "${HOMEBREW_PREFIX}/bin/codium-insiders"
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/VSCodium - Insiders.app/Contents/Resources/app/bin/codium-insiders' "${HOMEBREW_PREFIX}/bin/codium-insiders"
   # if we are using 'vscodium-insiders' only, symlink it to 'codium' for ease of typing
   replace_executable_if_exists_and_is_not_symlinked "${HOMEBREW_PREFIX}/bin/codium-insiders" "${HOMEBREW_PREFIX}/bin/codium"
   # extra: also symlink for 'code'
   replace_executable_if_exists_and_is_not_symlinked "${HOMEBREW_PREFIX}/bin/codium" "${HOMEBREW_PREFIX}/bin/code"
-elif is_directory "/Applications/VSCodium.app"; then
+  success 'Successfully linked vscodium-insiders into PATH'
+elif is_directory '/Applications/VSCodium.app'; then
   # Symlink from the embedded executable for codium
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/VSCodium.app/Contents/Resources/app/bin/codium" "${HOMEBREW_PREFIX}/bin/codium"
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/VSCodium.app/Contents/Resources/app/bin/codium' "${HOMEBREW_PREFIX}/bin/codium"
   # extra: also symlink for 'code'
   replace_executable_if_exists_and_is_not_symlinked "${HOMEBREW_PREFIX}/bin/codium" "${HOMEBREW_PREFIX}/bin/code"
-elif is_directory "/Applications/VSCode.app"; then
+  success 'Successfully linked vscodium into PATH'
+elif is_directory '/Applications/VSCode.app'; then
   # Symlink from the embedded executable for code
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/VSCode.app/Contents/Resources/app/bin/code" "${HOMEBREW_PREFIX}/bin/code"
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/VSCode.app/Contents/Resources/app/bin/code' "${HOMEBREW_PREFIX}/bin/code"
+  success 'Successfully linked vscode into PATH'
 else
-  warn "skipping symlinking vscode/vscodium for command-line invocation"
+  warn 'skipping symlinking vscode/vscodium for command-line invocation'
 fi
 
-section_header "Linking rider for command-line invocation"
-if is_directory "/Applications/Rider.app"; then
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/Rider.app/Contents/MacOS/rider" "${HOMEBREW_PREFIX}/bin/rider"
+section_header 'Linking rider for command-line invocation'
+if is_directory '/Applications/Rider.app'; then
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/Rider.app/Contents/MacOS/rider' "${HOMEBREW_PREFIX}/bin/rider"
+  success 'Successfully linked rider into PATH'
 else
-  warn "skipping symlinking rider for command-line invocation"
+  warn 'skipping symlinking rider for command-line invocation'
 fi
 
-section_header "Linking idea/idea-ce for command-line invocation"
-if is_directory "/Applications/IntelliJ IDEA CE.app"; then
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/IntelliJ IDEA CE.app/Contents/MacOS/idea" "${HOMEBREW_PREFIX}/bin/idea"
-elif is_directory "/Applications/IntelliJ IDEA.app"; then
-  replace_executable_if_exists_and_is_not_symlinked "/Applications/IntelliJ IDEA.app/Contents/MacOS/idea" "${HOMEBREW_PREFIX}/bin/idea"
+section_header 'Linking idea/idea-ce for command-line invocation'
+if is_directory '/Applications/IntelliJ IDEA CE.app'; then
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/IntelliJ IDEA CE.app/Contents/MacOS/idea' "${HOMEBREW_PREFIX}/bin/idea"
+  success 'Successfully linked idea-ce into PATH'
+elif is_directory '/Applications/IntelliJ IDEA.app'; then
+  replace_executable_if_exists_and_is_not_symlinked '/Applications/IntelliJ IDEA.app/Contents/MacOS/idea' "${HOMEBREW_PREFIX}/bin/idea"
+  success 'Successfully linked idea into PATH'
 else
-  warn "skipping symlinking idea/idea-ce for command-line invocation"
+  warn 'skipping symlinking idea/idea-ce for command-line invocation'
 fi
 
 #####################
 # Setup login items #
 #####################
-section_header "Setting up login items"
+section_header 'Setting up login items'
 setup_login_item() {
   if is_directory "/Applications/${1}"; then
-    echo "Setting up '${1}' as a login item" && osascript -e "tell application \"System Events\" to make login item at end with properties {path:\"/Applications/${1}\", hidden:false}" 2>&1 > /dev/null
+    osascript -e "tell application \"System Events\" to make login item at end with properties {path:\"/Applications/${1}\", hidden:false}" 2>&1 > /dev/null && success "Successfully setup '$(yellow "${1}")' $(green "as a login item")"
   else
     warn "Couldn't find application '/Applications/${1}' and so skipping setting up as a login item"
   fi
@@ -255,6 +268,7 @@ app_list=(
   'Itsycal.app'
   'KeepingYouAwake.app'
   'Keybase.app'
+  'KeyCastr.app'
   'Raycast.app'
   'Stats.app'
   'ZoomHider.app'
@@ -264,4 +278,4 @@ for app in "${app_list[@]}"; do
 done
 
 echo "\n"
-success "** Finished auto installation process: MANUALLY QUIT AND RESTART iTerm2 and Terminal apps **"
+success '** Finished auto installation process: MANUALLY QUIT AND RESTART iTerm2 and Terminal apps **'